+++ src/TQVaultAE.GUI/Components/PlayerPanel.cs
@@ -3,13 +3,14 @@
 //     Copyright (c) Brandon Wallace and Jesse Calhoun. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Components
 {
 	using Properties;
 	using System;
 	using System.Drawing;
 	using System.Windows.Forms;
 	using Tooltip;
+	using TQVaultAE.GUI.Models;
 	using TQVaultData;
 
 	/// <summary>
