+++ src/TQVaultAE.GUI/Components/SackPanel.cs
@@ -3,7 +3,7 @@
 //     Copyright (c) Brandon Wallace and Jesse Calhoun. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Components
 {
 	using Properties;
 	using System;
@@ -14,6 +14,7 @@ namespace TQVaultAE.GUI
 	using System.IO;
 	using System.Linq;
 	using System.Windows.Forms;
+	using TQVaultAE.GUI.Models;
 	using TQVaultData;
 
 	/// <summary>
