+++ src/TQVaultAE.GUI/Components/ScalingButton.resx
