+++ src/TQVaultAE.GUI/Components/ScalingCheckBox.cs
@@ -3,7 +3,7 @@
 //     Copyright (c) Brandon Wallace and Jesse Calhoun. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Components
 {
 	using System.Drawing;
 	using System.Windows.Forms;
