+++ src/TQVaultAE.GUI/Components/ScalingComboBox.cs
@@ -1,18 +1,32 @@
 ﻿//-----------------------------------------------------------------------
-// <copyright file="ScalingRadioButton.cs" company="None">
+// <copyright file="ScalingComboBox.cs" company="None">
 //     Copyright (c) Brandon Wallace and Jesse Calhoun. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Components
 {
 	using System.Drawing;
 	using System.Windows.Forms;
 
 	/// <summary>
-	/// RadioButton class to support scaling of the fonts.
+	/// ComboBox class to support scaling of the fonts.
 	/// </summary>
-	public class ScalingRadioButton : RadioButton
+	public class ScalingComboBox : ComboBox
 	{
+		/// <summary>
+		/// Reverts the basic settings of a control back to the original settings.
+		/// </summary>
+		/// <param name="location">New Location of the control</param>
+		/// <param name="size">New Size of the control</param>
+		public void Revert(Point location, Size size)
+		{
+			this.Font = Program.GetFontAlbertusMTLight(8.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+			this.Location = location;
+			this.Size = size;
+			this.BackColor = SystemColors.Window;
+			this.ForeColor = SystemColors.WindowText;
+		}
+
 		/// <summary>
 		/// Override of ScaleControl which supports font scaling.
 		/// </summary>
