+++ src/TQVaultAE.GUI/Components/StashPanel.cs
@@ -3,7 +3,7 @@
 //     Copyright (c) Brandon Wallace and Jesse Calhoun. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Components
 {
 	using Properties;
 	using System;
@@ -12,6 +12,7 @@ namespace TQVaultAE.GUI
 	using System.Text;
 	using System.Windows.Forms;
 	using Tooltip;
+	using TQVaultAE.GUI.Models;
 	using TQVaultData;
 
 	/// <summary>
