+++ src/TQVaultAE.GUI/Models/ComboBoxItem.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Models
 {
 	internal class ComboBoxItem
 	{
