+++ src/TQVaultAE.GUI/Models/SackPanelEventArgs.cs
@@ -3,7 +3,7 @@
 //     Copyright (c) Brandon Wallace and Jesse Calhoun. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Models
 {
 	using System;
 	using TQVaultData;
