+++ src/TQVaultAE.GUI/Models/User32.cs
@@ -3,7 +3,7 @@
 //     Copyright (c) Brandon Wallace and Jesse Calhoun. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
-namespace TQVaultAE.GUI
+namespace TQVaultAE.GUI.Models
 {
 	/// <summary>
 	/// Class encapsulating the User messages.
